# backend/Code/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_import_document_api_import_document_post": {
        "properties": {
          "component_subject_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Component Subject Id"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "section": {
            "title": "Section",
            "type": "string"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          }
        },
        "required": [
          "file",
          "subject_id",
          "section",
          "language"
        ],
        "title": "Body_import_document_api_import_document_post",
        "type": "object"
      },
...
      "DocumentImportResponse": {
        "additionalProperties": false,
        "description": "API-shaped result of one parsed and indexed upload.",
        "properties": {
          "added_at": {
            "format": "date-time",
            "title": "Added At",
            "type": "string"
          },
          "chunks_indexed": {
            "minimum": 0.0,
            "title": "Chunks Indexed",
            "type": "integer"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "material_id": {
            "title": "Material Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "pages": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pages"
          },
          "section": {
            "title": "Section",
            "type": "string"
          },
          "status": {
            "enum": [
              "indexed",
              "needs_review"
            ],
            "title": "Status",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "material_id",
          "name",
          "type",
          "section",
          "language",
          "chunks_indexed",
          "status",
          "added_at"
        ],
        "title": "DocumentImportResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LearningGoalModel": {
        "additionalProperties": false,
        "description": "Structured learning goal owned by the Python backend.",
        "properties": {
          "component_subject_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Component Subject Id"
          },
          "learning_goal_id": {
            "title": "Learning Goal Id",
            "type": "string"
          },
          "order": {
            "default": 0,
            "minimum": 0.0,
            "title": "Order",
            "type": "integer"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MaterialSourceModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "learning_goal_id",
          "subject_id",
          "title",
          "text"
        ],
        "title": "LearningGoalModel",
        "type": "object"
      },
      "MaterialModel": {
        "additionalProperties": false,
        "description": "One uploaded or canonical material item and its index state.",
        "properties": {
          "archived": {
            "default": false,
            "title": "Archived",
            "type": "boolean"
          },
          "chunks_indexed": {
            "default": 0,
            "minimum": 0.0,
            "title": "Chunks Indexed",
            "type": "integer"
          },
          "component_subject_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Component Subject Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "file_name": {
            "title": "File Name",
            "type": "string"
          },
          "indexed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Indexed At"
          },
          "material_id": {
            "title": "Material Id",
            "type": "string"
          },
          "material_name": {
            "title": "Material Name",
            "type": "string"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "section": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section"
          },
          "size_bytes": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/MaterialSourceModel"
            },
            "title": "Sources",
            "type": "array"
          },
          "status": {
            "enum": [
              "uploaded",
              "indexing",
              "indexed",
              "failed",
              "needs_review"
            ],
            "title": "Status",
            "type": "string"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          },
          "warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warning"
          }
        },
        "required": [
          "material_id",
          "subject_id",
          "material_name",
          "file_name",
          "media_type",
          "status",
          "created_at"
        ],
        "title": "MaterialModel",
        "type": "object"
      },
      "MaterialSourceModel": {
        "additionalProperties": false,
        "description": "Citation-oriented metadata for material content.",
        "properties": {
          "modality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modality"
          },
          "page": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_layer": {
            "title": "Source Layer",
            "type": "string"
          },
          "source_name": {
            "title": "Source Name",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "source_id",
          "source_name",
          "source_layer"
        ],
        "title": "MaterialSourceModel",
        "type": "object"
      },
...
      "SubjectComponentModel": {
        "additionalProperties": false,
        "description": "Nested concrete component of the virtual SPF subject.",
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "language": {
            "enum": [
              "de",
              "en",
              "fr"
            ],
            "title": "Language",
            "type": "string"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          }
        },
        "required": [
          "subject_id",
          "display_name",
          "language"
        ],
        "title": "SubjectComponentModel",
        "type": "object"
      },
      "SubjectModel": {
        "additionalProperties": false,
        "description": "One top-level subject exposed to the frontend.",
        "properties": {
          "components": {
            "items": {
              "$ref": "#/components/schemas/SubjectComponentModel"
            },
            "title": "Components",
            "type": "array"
          },
          "corpus_keys": {
            "items": {
              "type": "string"
            },
            "title": "Corpus Keys",
            "type": "array"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "indexed_materials": {
            "default": 0,
            "minimum": 0.0,
            "title": "Indexed Materials",
            "type": "integer"
          },
          "is_virtual": {
            "default": false,
            "title": "Is Virtual",
            "type": "boolean"
          },
          "language": {
            "enum": [
              "de",
              "en",
              "fr"
            ],
            "title": "Language",
            "type": "string"
          },
          "last_indexed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Indexed At"
          },
          "learning_goal_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Learning Goal Count",
            "type": "integer"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          },
          "topics": {
            "items": {
              "type": "string"
            },
            "title": "Topics",
            "type": "array"
          }
        },
        "required": [
          "subject_id",
          "display_name",
          "language"
        ],
        "title": "SubjectModel",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local Python AI/RAG backend for the Lovable frontend.",
    "title": "Alim Study Assistant API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/import/document": {
      "post": {
        "description": "Stream, parse, and rebuild only the identified student's corpus.",
        "operationId": "import_document_api_import_document_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Student-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Student-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_document_api_import_document_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Document",
        "tags": [
          "materials"
        ]
      }
    },
...
    "/api/subjects": {
      "get": {
        "description": "Return exactly 15 top-level subjects; omit private metadata when anonymous.",
        "operationId": "list_subject_catalog_api_subjects_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Student-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Student-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SubjectModel"
                  },
                  "title": "Response List Subject Catalog Api Subjects Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Subject Catalog",
        "tags": [
          "subjects"
        ]
      }
    },
    "/api/subjects/{subject_id}": {
      "get": {
        "description": "Return aggregate or component-specific corpus metadata.",
        "operationId": "get_subject_detail_api_subjects__subject_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "subject_id",
            "required": true,
            "schema": {
              "title": "Subject Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "component_subject_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Component Subject Id"
            }
          },
          {
            "in": "header",
            "name": "X-Student-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Student-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Subject Detail",
        "tags": [
          "subjects"
        ]
      }
    },
    "/api/subjects/{subject_id}/learning-goals": {
      "get": {
        "description": "Return user-specific goals when identified, otherwise canonical goals.",
        "operationId": "get_learning_goals_api_subjects__subject_id__learning_goals_get",
        "parameters": [
          {
            "in": "path",
            "name": "subject_id",
            "required": true,
            "schema": {
              "title": "Subject Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "component_subject_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Component Subject Id"
            }
          },
          {
            "in": "query",
            "name": "topic",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topic"
            }
          },
          {
            "in": "header",
            "name": "X-Student-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Student-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LearningGoalModel"
                  },
                  "title": "Response Get Learning Goals Api Subjects  Subject Id  Learning Goals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Learning Goals",
        "tags": [
          "subjects"
        ]
      }
    },
    "/api/subjects/{subject_id}/materials": {
      "get": {
        "description": "Return canonical-only material anonymously or isolated user material with identity.",
        "operationId": "get_materials_api_subjects__subject_id__materials_get",
        "parameters": [
          {
            "in": "path",
            "name": "subject_id",
            "required": true,
            "schema": {
              "title": "Subject Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "component_subject_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Component Subject Id"
            }
          },
          {
            "in": "query",
            "name": "section",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Section"
            }
          },
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "X-Student-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Student-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MaterialModel"
                  },
                  "title": "Response Get Materials Api Subjects  Subject Id  Materials Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Materials",
        "tags": [
          "subjects"
        ]
      }
    },
...
  }
}